Move shared types out of utilities file and into types file
